Add feature button in navbar
