+++ src/components/hero/hero.tsx
@@ -5,98 +5,36 @@ import {
   DropdownItem,
   DropdownMenu,
   DropdownTrigger,
+  Image,
 } from "@nextui-org/react";
 
-import {
-  ChevronDown,
-  Flash,
-  Scale,
-  Server,
-  Lock,
-  Activity,
-  TagUser,
-} from "./icons";
-
 const HeroComponent = () => {
-  const icons = {
-    chevron: <ChevronDown fill="currentColor" size={16} />,
-    scale: <Scale className="text-warning" fill="currentColor" size={30} />,
-    lock: <Lock className="text-success" fill="currentColor" size={30} />,
-    activity: (
-      <Activity className="text-secondary" fill="currentColor" size={30} />
-    ),
-    flash: <Flash className="text-primary" fill="currentColor" size={30} />,
-    server: <Server className="text-success" fill="currentColor" size={30} />,
-    user: <TagUser className="text-danger" fill="currentColor" size={30} />,
-  };
-
   return (
     <motion.div
       initial={{ opacity: 0, y: 100 }}
       animate={{ opacity: 1, y: 0 }}
-      transition={{ duration: 0.5, delay: 0.1, type: "spring" }}
+      transition={{ duration: 0.5, delay: 0.4, type: "spring" }}
       className="relative w-full mx-auto my-4"
     >
-      <div className="flex justify-center">
-        <Button color="primary" variant="flat">
-          <Dropdown>
-            <DropdownTrigger>
-              <Button
-                disableRipple
-                className="p-0 bg-transparent data-[hover=true]:bg-transparent"
-                endContent={icons.chevron}
-                radius="sm"
-                variant="light"
-              >
-                Features
-              </Button>
-            </DropdownTrigger>
-            <DropdownMenu
-              aria-label="ACME features"
-              className="w-[340px]"
-              itemClasses={{
-                base: "gap-4",
-              }}
-            >
-              <DropdownItem
-                key="autoscaling"
-                description="ACME scales apps to meet user demand, automatically, based on load."
-                startContent={icons.scale}
-              >
-                Autoscaling
-              </DropdownItem>
-              <DropdownItem
-                key="usage_metrics"
-                description="Real-time metrics to debug issues. Slow query added? We’ll show you exactly where."
-                startContent={icons.activity}
-              >
-                Usage Metrics
-              </DropdownItem>
-              <DropdownItem
-                key="production_ready"
-                description="ACME runs on ACME, join us and others serving requests at web scale."
-                startContent={icons.flash}
-              >
-                Production Ready
-              </DropdownItem>
-              <DropdownItem
-                key="99_uptime"
-                description="Applications stay on the grid with high availability and high uptime guarantees."
-                startContent={icons.server}
-              >
-                +99% Uptime
-              </DropdownItem>
-              <DropdownItem
-                key="supreme_support"
-                description="Overcome any challenge with a supporting team ready to respond."
-                startContent={icons.user}
-              >
-                +Supreme Support
-              </DropdownItem>
-            </DropdownMenu>
-          </Dropdown>
-        </Button>
-      </div>
+      <motion.div
+        initial={{ opacity: 0, y: 100 }}
+        animate={{ opacity: 1, y: 0 }}
+        transition={{ duration: 0.5, delay: 0.1, type: "spring" }}
+        className="grid grid-cols-2 place-content-around gap-4 m-5 p-8"
+      >
+        <div className="flex justify-center">
+          <div>Code Odyssey Editor</div>
+        </div>
+        <div className="flex justify-center mr-10">
+          <Image
+            isBlurred
+            width={400}
+            src="/logo/logo.png"
+            alt="NextUI Album Cover"
+            className="m-5 hidden sm:block"
+          />
+        </div>
+      </motion.div>
     </motion.div>
   );
 };
