+++ src/components/navbar/navbar.tsx
@@ -10,11 +10,36 @@ import {
   NavbarMenuToggle,
   NavbarMenu,
   NavbarMenuItem,
+  Dropdown,
+  DropdownTrigger,
+  DropdownMenu,
+  DropdownItem,
 } from "@nextui-org/react";
 import { AcmeLogo } from "./AcmeLogo.jsx";
 import ThemeSwitcher from "../theme_switcher/theme_toggle.tsx";
+import {
+  ChevronDown,
+  Scale,
+  Lock,
+  Activity,
+  Flash,
+  Server,
+  TagUser,
+} from "./icons.tsx";
 
 export default function NavbarComponent() {
+  const icons = {
+    chevron: <ChevronDown fill="currentColor" size={16} />,
+    scale: <Scale className="text-warning" fill="currentColor" size={30} />,
+    lock: <Lock className="text-success" fill="currentColor" size={30} />,
+    activity: (
+      <Activity className="text-secondary" fill="currentColor" size={30} />
+    ),
+    flash: <Flash className="text-primary" fill="currentColor" size={30} />,
+    server: <Server className="text-success" fill="currentColor" size={30} />,
+    user: <TagUser className="text-danger" fill="currentColor" size={30} />,
+  };
+
   const [isMenuOpen, setIsMenuOpen] = React.useState(false);
 
   const menuItems = [
@@ -32,8 +57,8 @@ export default function NavbarComponent() {
 
   return (
     <Navbar
-    position="static"
-    onMenuOpenChange={setIsMenuOpen}
+      position="static"
+      onMenuOpenChange={setIsMenuOpen}
       className="backdrop:blur rounded-full bg-pink-100/20 border border-gray-100/40 dark:border-gray-800/40 dark:bg-gray-800/80 w-3/4 mx-auto mt-4"
     >
       <NavbarContent>
@@ -41,16 +66,69 @@ export default function NavbarComponent() {
           aria-label={isMenuOpen ? "Close menu" : "Open menu"}
           className="md:hidden text-black dark:text-white/80"
         />
-      <NavbarBrand>
-        <AcmeLogo />
-        <p className="font-bold text-inherit">Code Odyssey Editor</p>
-      </NavbarBrand>
+        <NavbarBrand>
+          <AcmeLogo />
+          <p className="font-bold text-inherit">Code Odyssey Editor</p>
+        </NavbarBrand>
       </NavbarContent>
       <NavbarContent className="hidden lg:flex gap-4" justify="center">
         <NavbarItem>
-          <Link color="foreground" href="#">
-            Features
-          </Link>
+          <Dropdown>
+            <DropdownTrigger>
+              <Button
+                disableRipple
+                className="p-0 bg-transparent data-[hover=true]:bg-transparent"
+                endContent={icons.chevron}
+                radius="sm"
+                variant="light"
+              >
+                Features
+              </Button>
+            </DropdownTrigger>
+            <DropdownMenu
+              aria-label="ACME features"
+              className="w-[340px]"
+              itemClasses={{
+                base: "gap-4",
+              }}
+            >
+              <DropdownItem
+                key="autoscaling"
+                description="ACME scales apps to meet user demand, automatically, based on load."
+                startContent={icons.scale}
+              >
+                Autoscaling
+              </DropdownItem>
+              <DropdownItem
+                key="usage_metrics"
+                description="Real-time metrics to debug issues. Slow query added? We’ll show you exactly where."
+                startContent={icons.activity}
+              >
+                Usage Metrics
+              </DropdownItem>
+              <DropdownItem
+                key="production_ready"
+                description="ACME runs on ACME, join us and others serving requests at web scale."
+                startContent={icons.flash}
+              >
+                Production Ready
+              </DropdownItem>
+              <DropdownItem
+                key="99_uptime"
+                description="Applications stay on the grid with high availability and high uptime guarantees."
+                startContent={icons.server}
+              >
+                +99% Uptime
+              </DropdownItem>
+              <DropdownItem
+                key="supreme_support"
+                description="Overcome any challenge with a supporting team ready to respond."
+                startContent={icons.user}
+              >
+                +Supreme Support
+              </DropdownItem>
+            </DropdownMenu>
+          </Dropdown>
         </NavbarItem>
         <NavbarItem isActive>
           <Link href="#" aria-current="page">
@@ -83,7 +161,11 @@ export default function NavbarComponent() {
           <NavbarMenuItem key={`${item}-${index}`}>
             <Link
               color={
-                index === 2 ? "primary" : index === menuItems.length - 1 ? "danger" : "foreground"
+                index === 2
+                  ? "primary"
+                  : index === menuItems.length - 1
+                  ? "danger"
+                  : "foreground"
               }
               className="w-full"
               href="#"
